+++ src/main/java/uk/gov/pay/connector/gatewayaccount/model/GatewayAccountRequest.java
@@ -139,72 +139,4 @@ public class GatewayAccountRequest {
     public boolean isAllowGooglePay() {
         return allowGooglePay;
     }
-    
-    public static final class Builder {
-        private String providerAccountType;
-        private String serviceName;
-        private String serviceId;
-        private String description;
-        private String analyticsId;
-        private String paymentProvider;
-        private boolean requires3ds;
-        private boolean allowApplePay;
-        private boolean allowGooglePay;
-
-        private Builder() {
-        }
-
-        public static Builder builder() {
-            return new Builder();
-        }
-
-        public Builder withProviderAccountType(String providerAccountType) {
-            this.providerAccountType = providerAccountType;
-            return this;
-        }
-
-        public Builder withServiceName(String serviceName) {
-            this.serviceName = serviceName;
-            return this;
-        }
-
-        public Builder withServiceId(String serviceId) {
-            this.serviceId = serviceId;
-            return this;
-        }
-
-        public Builder withDescription(String description) {
-            this.description = description;
-            return this;
-        }
-
-        public Builder withAnalyticsId(String analyticsId) {
-            this.analyticsId = analyticsId;
-            return this;
-        }
-
-        public Builder withPaymentProvider(String paymentProvider) {
-            this.paymentProvider = paymentProvider;
-            return this;
-        }
-
-        public Builder withRequires3ds(boolean requires3ds) {
-            this.requires3ds = requires3ds;
-            return this;
-        }
-
-        public Builder withAllowApplePay(boolean allowApplePay) {
-            this.allowApplePay = allowApplePay;
-            return this;
-        }
-
-        public Builder withAllowGooglePay(boolean allowGooglePay) {
-            this.allowGooglePay = allowGooglePay;
-            return this;
-        }
-
-        public GatewayAccountRequest build() {
-            return new GatewayAccountRequest(providerAccountType, paymentProvider, serviceName, serviceId, description, analyticsId, requires3ds, allowApplePay, allowGooglePay);
-        }
-    }
 }
